broke things out into separate daoProject, daoUser, etc files - also did additional work creating more useful queries
