+++ portfolio-project-matching-app/Firebase/clientApp.ts
@@ -82,12 +82,14 @@ const addNewDoc = async (collectionName, data) => {
     DESCRIPTION:  adds a new document with provided data to the specified
                   collection. This function does not perform any input
                   validation - that is currently left to the calling function.
+                  Note, document ID is generated by Firebase
 
     INPUT:        string indicating desired collection name and data to be
                   added. E.g.,
 
                   addNewDoc('technologies', {name: _, language: _, ...})
-    RETURN:       NA
+
+    RETURN:       reference to new document
   */
   const coll = collection(db, collectionName);
   const docRef = await addDoc(coll, data);
@@ -95,7 +97,33 @@ const addNewDoc = async (collectionName, data) => {
 }
 
 const addNewDocWithId = async (collectionName, id, data) => {
-  const docRef = setDoc(doc(db, collectionName, id), data);
+  /*
+    DESCRIPTION:  adds a new document with provided data to the specified
+                  collection. This function does not perform any input
+                  validation - that is currently left to the calling function.
+                  Note, new document id must be provided by calling function
+
+    INPUT:        string indicating desired collection name and id of new
+                  document
+
+    RETURN:       reference to new document
+  */
+  const docRef = doc(db, collectionName, id);
+  await setDoc(docRef, data);
+  return docRef;
+}
+
+const deleteDocById = async (collectionName, id) => {
+  /*
+    DESCRIPTION:  deletes document with specified id from specified collection
+
+    INPUT:        string indicating desired collection name and id of document
+                  to be deleted
+
+    RETURN:       todo
+  */
+  const docRef = doc(db, collectionName, id);
+  await deleteDoc(docRef);
   return docRef;
 }
 
@@ -133,11 +161,12 @@ const addNewDocWithId = async (collectionName, id, data) => {
 
 // export
 export {
-  db,
-  firebaseApp,
-  getCollectionSnapshot,
   addNewDoc,
   addNewDocWithId,
+  db,
+  deleteDocById,
+  firebaseApp,
+  getCollectionSnapshot,
+  getCollectionSnapshotByCriteria,
   getDocSnapshotById,
-  getCollectionSnapshotByCriteria
 };
